proto: add helper methods for working with proto types
Split into logical files:
- instance.go: InstanceIDHex() for StartInstance, Vote, Decided
- request.go: TransactionsByChain() for XTRequest

# Conflicts:
#	compose/proto/instance.go
#	compose/proto/instance_test.go
