+++ compose/proto/instance_test.go
@@ -29,22 +29,3 @@ func TestInstanceIDHex_Nil(t *testing.T) {
 	var d *Decided
 	assert.Empty(t, d.InstanceIDHex())
 }
-
-func TestTransactionsByChain(t *testing.T) {
-	req := &XTRequest{
-		TransactionRequests: []*TransactionRequest{
-			{ChainId: 901, Transaction: [][]byte{{0x01}, {0x02}}},
-			{ChainId: 902, Transaction: [][]byte{{0x03}}},
-		},
-	}
-
-	m := req.TransactionsByChain()
-	assert.Len(t, m, 2)
-	assert.Equal(t, [][]byte{{0x01}, {0x02}}, m[901])
-	assert.Equal(t, [][]byte{{0x03}}, m[902])
-}
-
-func TestTransactionsByChain_Nil(t *testing.T) {
-	var req *XTRequest
-	assert.Nil(t, req.TransactionsByChain())
-}
